Item View Item Controller の整備

+++ app/Http/Controllers/Admin/Item/DeleteController.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace App\Http\Contrrollers\Address;
+namespace App\Http\Contrrollers\Admin\Item;
 
+use App\Item;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 
-class HomeController extends Controller
+class DeleteController extends Controller
 {
     /**
      * Create a new controller instance.
@@ -24,10 +26,15 @@ class HomeController extends Controller
 	}
 
 
-    public function delete($id)
+    public function delete(Request $request)
     {
-		$data = Item::findOrFail($id);
-		$data->delete();
-		return redirect('/Item/home');
+		if (isset($request->id)) {
+			$data = Item::findOrFail($id);
+			$data->delete();
+			return view('Item.index');
+		} else {
+			$items = Item::all();
+			return view('Item.index', ['items' => $items]);
+		}
     }
 }
